feat: ビュー統計情報のキャッシュ機能を追加し、StatsManagerからのキャッシュ処理をCacheManagerに統合
